requirements.txt
accelerate==1.10.1
ace_tools==0.0
addict==2.4.0
airportsdata==20250811
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
av==15.1.0
blinker==1.9.0
bypy==1.8.9
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.3
crcmod==1.7
cryptography==45.0.6
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
defusedxml==0.7.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
einops==0.8.1
executing==2.2.0
fastapi==0.116.1
filelock==3.14.0
fire==0.7.1
fonttools==4.59.0
fsspec==2025.5.1
genson==1.3.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
interegular==0.3.3
iopath==0.1.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==0.10.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lap==0.5.12
lark==1.2.2
lmdeploy==0.9.2.post1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mmengine-lite==0.10.7
model-index==0.1.11
mpmath==1.3.0
msgpack==1.1.1
multiprocess==0.70.18
narwhals==2.0.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.104.2
opencv-python==4.12.0.88
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
outlines==0.1.14
outlines_core==0.1.26
packaging==24.2
pandas==2.3.0
parso==0.8.4
partial-json-parser==0.2.1.1.post6
peft==0.14.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
portalocker==3.2.0
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML==6.0.2
pyzmq==27.0.0
qwen-vl-utils==0.0.11
ray==2.49.1
referencing==0.36.2
regex==2025.9.1
requests==2.28.2
requests-toolbelt==1.0.0
rich==13.4.2
rpds-py==0.26.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
sentencepiece==0.2.1
setuptools==80.9.0
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.3
streamlit==1.47.1
supervision==0.26.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
toml==0.10.2
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tornado==6.5.1
tqdm==4.65.2
traitlets==5.14.3
# Editable install with no version control (transformers==4.57.0.dev0)
-e /home/jinqiao/mhi/transformers
triton==3.2.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.170
ultralytics-thop==2.0.14
urllib3==1.26.20
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
wheel==0.45.1
yapf==0.43.0
yt-dlp==2025.6.9
